[buildarmimg] change the logic for partitioning Lately we have started to support some "special" devices, like the Quartz64-BSP or Generic devices. These require special setups in partitioning for various reasons. This resulting in a mess of if statements all through the create_img function.
So change the partitioning logic to something more "logical". It's more code, but should be easier to follow and understand as special devices now get a complete case statement instead of scattered if statements in the function.

Also, increase the boot partition end to 512 MB to prepare for possible multiple kernel setups better.

Signed-off-by: Dan Johansen <strit@manjaro.org>