[buildarmimg] download and flash pbpro-bsp uboot directly There have been reports of people having messed up their Pinebook Pro's still running the factory install, by updating the uboot package we ship on it and flashin it.
So to avoid that, we now download and flash the binaries directly, when we make a pbpro-bsp image instead of using a package.

This makes the package uboot-pinebookpro-bsp obsolete and should be removed from the pbpro-bsp profile and then from the repository as well.
I will do this later. Only our github CI should be creating images for this profile.

Existing users of the package, will get switched over to the mainline uboot-pinebookpro package later, as the BSP is only useful for testing at the factory at this point.

Signed-off-by: Dan Johansen <strit@manjaro.org>