Various improvements
Perform various small cleanups, as spotted, which include
proper indentation, uniform style of function and variable
names, some improvements to the produced help messages and
comments in the code, removal of redundant parameters for
the "cp -a" invocations, removal of the redundant trailing
slashes, simplification of redundant "1>" redirections,
small code readability improvements, and quite a bit of
other small coding style cleanups.

Perform a rewrite of the entire logic that handles local
packages that are to be installed to the root filesystem
in the images.  Previously, that logic was overly complex
and somewhat broken, more specifically it performed various
redundants operations, and it did not work under certain
conditions because it expected the local packages to be
actually present in the pacman cache, etc.

At the same time, call the verify_local_packages() function
in "bin/buildemmcinstaller" as well, to have the same check
as the one that's already performed in the "bin/buildarmimg"
and "bin/buildarmpkg" files.

Signed-off-by: Dragan Simic <dsimic@manjaro.org>