More fixes and cleanups
The last commit or two failed to resolve the issues with bind
mounting directories with the local packages, because the current
directory was changed somewhere and the directory with the local
packages no longer matched the real location.  Fix that by saving
the actual directory early and reusing it later.  Also, change
a few invocations of "wget" so no current directory changes are
performed prior to those invocations, in an atttempt to prevent
similar issues from happening again somewhere else.

At the same time, use the least amount of "force" while removing
empty temporary directories used as mount points, i.e. use "rmdir"
to remove them, instead of using "rm -rf".

Perform a few more small cleanups and improvements in the produced
messages and comments, as well as a bit of code deduplication in
error handling, as spotted.  Also, remove redundant "-N" parameters
in a few invocations of "wget".

Signed-off-by: Dragan Simic <dsimic@manjaro.org>