More bugfixes
Hopefully, fix the issues with installing local packages by
saving the list of package files in the check_local_pkgs()
function and using it later, to avoid issues with the wildcard
expansion in "pacman -U" invocations.

At the same time, simplify the handling of the absolute path
in the check_local_pkgs() function a bit.

Signed-off-by: Dragan Simic <dsimic@manjaro.org>