Improve the "rmdir canaries"
A few instances of executing "rm -r" or "rm -rf" on directories
have been changed to "rmdir" in previous commits, because those
directories should be empty at that point, which introduced some
kind of a "canary" for such directories.

Complete those "canaries" by actually aborting the execution of
a utility if such "rmdir" operations fail.  It's a critical error
anyway, because it means something went seriously wrong, and it
should be inspected by hand.

Signed-off-by: Dragan Simic <dsimic@manjaro.org>