[buildarmimg] add zstd image compression support
Currently all images are compressed using xz, which is really good compression,
butit also takes ages to compress and decompress.

So add the option to define zstd compression instead with the -z flag.
Compression and decompression time is greatly improved with zstd,
but the files also get a bit larger in the end.
We use compression level 10 as a good middle ground between time it takes
and filesize.

Signed-off-by: Dan Johansen <strit@manjaro.org>