Update functions.sh
For some reason the check did not work as expected. 
(possilby because I forgot to negate the second condition.
It should only exit if x86 was found but not the qemu-aarch64-static binary.
Rewritten to use a `if then fi` block to ensure the error message is only thrown 
if the combined expression evaluates to true